Remove variable costs of storage
These costs are speculative and potentially imprecise, especially for batteries. Simple aging of storage technologies can be modelled with the lifetime parameter.

Furthermore, removing variable costs allowed to scale costs by another factor of 100, which decreased the numeric range of the matrix (from 6 to 4 orders of magnitude) and decreases the objective value by 2 orders of magnitude. Both should be beneficial for the solver (barrier).