Fix suboptimal uncertainty runs
The numerical range of the RHS was pretty bad: because of the hydro energy time series, it went down to 1e-13. This led to suboptimal solutions by barrier.

Now, hydro uses capacity factor time series which are capped, so that they do not become too small. This seems to have solved the issue.