Remove tiny variable costs of batteries once again
I addded them in one of the last commits as I thought they were >= 0.1 ct/kWh. Instead, they are 0.0378 ct/kWh. This caused numerical issues in the uncertainty runs. Setting it to 0 made the uncertainty runs work. Anyways, those costs are really tiny and do not change the results significantly.