Model grid restriction using existing lines without capacity
Grid restrictions used to be modelled by literally removing the lines.
Now, I keep the lines existing, but I restrict their capacity to 0.

This has the advantage that import constraints don't fail --- which they do when no line exists. That avoids special case treatment.

It has the disadvantage that the lines indeed do exist. This may (or may not) be worse for solution performance. It's slightly worse for analysis as now it's more difficult to differentiate between a normal line and one that does not exist.