Fix cost normalised by demand
I was not strict enough with the terms cost and lcoe. Mostly, I am interested in total cost which is no problem. Sometimes I need to consider total cost normed by system size. So far I used the LCOE here, but that is not a good idea because different cases generate different amounts of electricity and thus, the quantity that I use for normalisation varies.

Now, I am avoiding to normalise cost whereever possible. When I do need to normalise, I use demand.