Remove om_prod from uncertainty analysis
The current way of modeling cost adds a 0.001 €/kWh penalty to wind and solar energy. This penalty is added for computational reasons only, and subtracted from annual cost in the deterministic model but not in the model with uncertainties.

There are two options here:
1) Model this correctly, which requires a lot of manual work.
2) Leave as is.

I decided for the second option, because the impact is small and the danger of introducing another error by fixing this manually is high.

What is the impact? Total cost may rise up to 0.001 €/kWh. Considering that total cost are > 0.05 €/kWh this is 2% or less and therefore very small.

Total cost is anyways not an important measure for our analysis. How does this model choice impact relative results on different geographic scales? This model choice adds a penalty to solar and wind, which means that energy from hydro and biomass are relatively cheaper. However, hydro is used to almost fullest extent always, as it is cheap in any case. Biomass, on the other hand, is very expensive in any case. Still, this choice may lead to slightly higher shares of biomass.